Artificial Neural Network
can run on GPU

There isn't much changes in the model but the accuracy computation. We need to change the pred and y tensors from GPU to CPU to do comparison.

In order to run the neural network model on GPU, we need to move the model and the train and test tensors to GPU, which are the only modifications.